feat: migrate to 4-tier soft labeling, add UI buttons, disable auto-balancer

+++ deepkt/db.py
@@ -58,7 +58,7 @@ def _init_tables(conn):
             id              INTEGER PRIMARY KEY AUTOINCREMENT,
             anchor_id       TEXT NOT NULL REFERENCES tracks(id),
             candidate_id    TEXT NOT NULL REFERENCES tracks(id),
-            label           INTEGER NOT NULL,
+            label           REAL NOT NULL,
             created_at      TIMESTAMP DEFAULT CURRENT_TIMESTAMP
         );
